feat: move create_suffixes to paths.unique_suffixes
